Simplify sleep logic
## What?

Attempts to simplify sleep logic by:

1. Organizing logs so they are more even-length and always happen after
   they are calculated.
2. Uses `saturating_sub` instead of `checked_sub` because it looks like
   std won't sleep if the `Duration` is zero anyway.
3. Using `split()` and `delta()` differently to avoid the overhead of
   resetting the clock in one spot and getting a more accurate sleep
   timing in another (though I'm unsure about this -- the subtraction
   still counts as time spent in the loop?)

## Why?

I felt like it was getting hard to read and wondered if this would work.
We originally used `checked_sub` to avoid the overhead of a syscall if
the sleep time was zero, but looking at std, we might not need to worry
about that. If we wanted to, we could extract our own function that
checks the `Duration` before doing the sleep.